chore(index.d.ts): update version to v0.12.19.0 in TypeScript definit…

+++ src/Moongate.UO.Data/Packets/Characters/PaperdollPacket.cs
@@ -30,9 +30,12 @@ public class PaperdollPacket : BaseUoPacket
             baseFlags |= 0x40;
         }
 
+
+
+
         baseFlags |= 0x02;
 
-        writer.Write(baseFlags);
+        writer.Write((byte)baseFlags);
 
 
         return writer.ToArray();
